+++ backend/InvestAndChill/src/main/java/com/starter/InvestAndChill/jwt/payload/response/PTC/Bal4Response.java
@@ -3,18 +3,18 @@ package com.starter.InvestAndChill.jwt.payload.response.PTC;
 import com.starter.InvestAndChill.jwt.payload.response.ReportReponse;
 
 public class Bal4Response extends ReportReponse{
-	private Integer NetDebt;
-	private Integer Workingcap;
-	public Integer getNetDebt() {
+	private Double NetDebt;
+	private Double Workingcap;
+	public Double getNetDebt() {
 		return NetDebt;
 	}
-	public void setNetDebt(Integer netDebt) {
+	public void setNetDebt(Double netDebt) {
 		NetDebt = netDebt;
 	}
-	public Integer getWorkingcap() {
+	public Double getWorkingcap() {
 		return Workingcap;
 	}
-	public void setWorkingcap(Integer workingcap) {
+	public void setWorkingcap(Double workingcap) {
 		Workingcap = workingcap;
 	}
