+++ backend/InvestAndChill/src/main/java/com/starter/InvestAndChill/jwt/payload/response/PTC/Perf2Response.java
@@ -3,32 +3,32 @@ package com.starter.InvestAndChill.jwt.payload.response.PTC;
 import com.starter.InvestAndChill.jwt.payload.response.ReportReponse;
 
 public class Perf2Response extends ReportReponse{
-	private Integer loiNhuanCotLoi;
-	private Integer loiNhuanTaiChinh;
-	private Integer ThuNhapKhac;
-	private Integer LaiLoTuCongTyLienDoanh;
-	public Integer getLoiNhuanCotLoi() {
+	private Double loiNhuanCotLoi;
+	private Double loiNhuanTaiChinh;
+	private Double ThuNhapKhac;
+	private Double LaiLoTuCongTyLienDoanh;
+	public Double getLoiNhuanCotLoi() {
 		return loiNhuanCotLoi;
 	}
-	public void setLoiNhuanCotLoi(Integer loiNhuanCotLoi) {
+	public void setLoiNhuanCotLoi(Double loiNhuanCotLoi) {
 		this.loiNhuanCotLoi = loiNhuanCotLoi;
 	}
-	public Integer getLoiNhuanTaiChinh() {
+	public Double getLoiNhuanTaiChinh() {
 		return loiNhuanTaiChinh;
 	}
-	public void setLoiNhuanTaiChinh(Integer loiNhuanTaiChinh) {
+	public void setLoiNhuanTaiChinh(Double loiNhuanTaiChinh) {
 		this.loiNhuanTaiChinh = loiNhuanTaiChinh;
 	}
-	public Integer getThuNhapKhac() {
+	public Double getThuNhapKhac() {
 		return ThuNhapKhac;
 	}
-	public void setThuNhapKhac(Integer thuNhapKhac) {
+	public void setThuNhapKhac(Double thuNhapKhac) {
 		ThuNhapKhac = thuNhapKhac;
 	}
-	public Integer getLaiLoTuCongTyLienDoanh() {
+	public Double getLaiLoTuCongTyLienDoanh() {
 		return LaiLoTuCongTyLienDoanh;
 	}
-	public void setLaiLoTuCongTyLienDoanh(Integer laiLoTuCongTyLienDoanh) {
+	public void setLaiLoTuCongTyLienDoanh(Double laiLoTuCongTyLienDoanh) {
 		LaiLoTuCongTyLienDoanh = laiLoTuCongTyLienDoanh;
 	}
